Deployment/requirements.txt
streamlit
transformers
torch
pillow
opencv-python
numpy
pytorch-grad-cam